# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "summary": "Get list of articles",
                "description": "Returns list of articles",
                "operationId": "getArticles",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Post a articles",
                "description": "Returns a articles",
                "operationId": "storeArticles",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/articles/{article}": {
            "get": {
                "summary": "Get detail of article",
                "description": "Returns a detail of article",
                "operationId": "showArticles",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "Post register user",
                "description": "Returns a user",
                "operationId": "postRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreArticleRequest": {
                "title": "Store Article request",
                "description": "Store Article request body data",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "title of the new article",
                        "type": "string",
                        "example": "title 01"
                    },
                    "body": {
                        "title": "body",
                        "description": "body of the new article",
                        "type": "string",
                        "example": "body 01"
                    }
                },
                "type": "object"
            },
            "StoreUserRequest": {
                "title": "Store User request",
                "description": "Store User request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new user",
                        "type": "string",
                        "example": "user123"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of the new user",
                        "type": "string",
                        "example": "user123@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of the new user",
                        "type": "string",
                        "format": "string",
                        "example": 123456
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "password_confirmation of the new user",
                        "type": "string",
                        "format": "string",
                        "example": 123456
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "401": {
                "description": "Unauthenticated"
            },
            "403": {
                "description": "Forbidden"
            }
        },
        "securitySchemes": {
            "api_key_security_example": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "api_token",
                "in": "query"
            }
        }
    },
    "security": [
        {
            "api_key_security_example": []
        }
    ]
}